# docs/openapi/gateway.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SatQuery Gateway",
    "description": "API Gateway and BFF for SatQuery AI",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/missions": {
      "get": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy List Missions",
        "description": "Proxy GET /api/v1/missions \u2192 Mission service.",
        "operationId": "proxy_list_missions_api_v1_missions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy Create Mission",
        "description": "Proxy POST /api/v1/missions \u2192 Mission service.",
        "operationId": "proxy_create_mission_api_v1_missions_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/missions/{mission_id}": {
      "get": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy Get Mission",
        "description": "Proxy GET /api/v1/missions/{id} \u2192 Mission service.",
        "operationId": "proxy_get_mission_api_v1_missions__mission_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy Update Mission",
        "description": "Proxy PATCH /api/v1/missions/{id} \u2192 Mission service.",
        "operationId": "proxy_update_mission_api_v1_missions__mission_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy Delete Mission",
        "description": "Proxy DELETE /api/v1/missions/{id} \u2192 Mission service (OPERATOR+).",
        "operationId": "proxy_delete_mission_api_v1_missions__mission_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/missions/{mission_id}/runs": {
      "post": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy Submit Run",
        "description": "Proxy POST /api/v1/missions/{id}/runs \u2192 Mission service (OPERATOR+).",
        "operationId": "proxy_submit_run_api_v1_missions__mission_id__runs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Mission Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy Get Job",
        "description": "Proxy GET /api/v1/jobs/{job_id} \u2192 Mission service.",
        "operationId": "proxy_get_job_api_v1_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/plan": {
      "post": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy Agent Plan",
        "description": "Proxy POST /api/v1/agent/plan \u2192 Agent service (ANALYST+).",
        "operationId": "proxy_agent_plan_api_v1_agent_plan_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/execute": {
      "post": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy Agent Execute",
        "description": "Proxy POST /api/v1/agent/execute \u2192 Agent service (ANALYST+).",
        "operationId": "proxy_agent_execute_api_v1_agent_execute_post",
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/runs/{job_id}": {
      "get": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy Agent Run Status",
        "description": "Proxy GET /api/v1/agent/runs/{job_id} \u2192 Agent service.",
        "operationId": "proxy_agent_run_status_api_v1_agent_runs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/sensor-decision": {
      "post": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy Agent Sensor Decision",
        "description": "Proxy POST /api/v1/agent/sensor-decision \u2192 Agent service.",
        "operationId": "proxy_agent_sensor_decision_api_v1_agent_sensor_decision_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/confidence": {
      "post": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy Agent Confidence",
        "description": "Proxy POST /api/v1/agent/confidence \u2192 Agent service.",
        "operationId": "proxy_agent_confidence_api_v1_agent_confidence_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/tools": {
      "get": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy Agent Tools",
        "description": "Proxy GET /api/v1/agent/tools \u2192 Agent service.",
        "operationId": "proxy_agent_tools_api_v1_agent_tools_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/inference/analyses": {
      "post": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy Inference Analyse",
        "description": "Proxy POST /api/v1/inference/analyses \u2192 Inference service (ANALYST+).\n\nReturns ``Analysis | Abstention`` unchanged. Both are 200: an abstention is a\nconsidered answer, not a transport failure.",
        "operationId": "proxy_inference_analyse_api_v1_inference_analyses_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/inference/models": {
      "get": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy Inference Models",
        "description": "Proxy GET /api/v1/inference/models \u2192 Inference service.\n\nThe registry: each model's held-out score, calibration, and whether its\ncheckpoint has been verified against its committed manifest.",
        "operationId": "proxy_inference_models_api_v1_inference_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/inference/analyses/{trace_id}/extent": {
      "get": {
        "tags": [
          "proxy"
        ],
        "summary": "Proxy Inference Extent",
        "description": "Proxy GET .../analyses/{trace_id}/extent \u2192 Inference service.\n\nGeoJSON, not JSON-wrapped: the map consumes it directly. ``trace_id`` is parsed\nas a UUID here as well as downstream, so the path forwarded is always one this\ngateway built -- A10 above -- never a caller's string.",
        "operationId": "proxy_inference_extent_api_v1_inference_analyses__trace_id__extent_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/health": {
      "get": {
        "summary": "Health Check",
        "operationId": "health_check_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthStatus"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthStatus": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "service": {
            "type": "string",
            "title": "Service"
          }
        },
        "type": "object",
        "required": [
          "status",
          "service"
        ],
        "title": "HealthStatus"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}