# services/gateway/routers/proxy.py
"""
services/gateway/routers/proxy.py — Reverse-proxy routes (P1-05).

Routes all /api/v1/missions/*, /api/v1/agent/* and /api/v1/inference/* calls to
the respective downstream microservices via the InternalClient (S2S auth,
circuit-breaker, retries, telemetry).

OWASP mitigations:
  A01 — Token verified by get_current_user before any proxy call.
  A07 — S2S token generated per-request; never logged.
  A10 — Server-Side Request Forgery prevented: downstream URLs are resolved from
        environment variables only (no user input in URL construction).
  A05 — Security misconfiguration: HTTPS enforced in prod by requiring HTTPS-
        prefixed base URLs; HTTP allowed only via explicit GATEWAY_ALLOW_HTTP env.
"""

import logging
import os
import uuid
from typing import Any, Dict, Optional

import httpx
from fastapi import APIRouter, Depends, HTTPException, Request
from fastapi.responses import JSONResponse, Response

from packages.auth.dependencies import require_role
from packages.auth.models import AuthContext, Role
from packages.shared.client import CircuitBreakerOpenError, InternalClient, InternalClientError

logger = logging.getLogger(__name__)

router = APIRouter(tags=["proxy"])

# ── Service base URLs ─────────────────────────────────────────────────────────
# Resolved once at import time; never constructed from user input.
_MISSION_URL = os.getenv("MISSION_SERVICE_URL", "http://localhost:8001")
_AGENT_URL = os.getenv("AGENT_SERVICE_URL", "http://localhost:8002")
_INFERENCE_URL = os.getenv("INFERENCE_SERVICE_URL", "http://localhost:8003")

#: One analysis reads, reprojects, runs the model and writes two artefacts. Measured
#: under a second for a 512x512 chip, but the shared client's 5 s default leaves no
#: room for a larger scene, and a timeout there is retried -- re-running inference.
_INFERENCE_TIMEOUT_S = float(os.getenv("INFERENCE_TIMEOUT_S", "30"))

# ── InternalClient singletons (lazy-init, reused across requests) ─────────────
_mission_client: Optional[InternalClient] = None
_agent_client: Optional[InternalClient] = None
_inference_client: Optional[InternalClient] = None


def _get_mission_client() -> InternalClient:
    global _mission_client
    if _mission_client is None:
        _mission_client = InternalClient(
            base_url=_MISSION_URL,
            caller_service="gateway",
            scopes=["mission:read", "mission:write"],
        )
    return _mission_client


def _get_agent_client() -> InternalClient:
    global _agent_client
    if _agent_client is None:
        _agent_client = InternalClient(
            base_url=_AGENT_URL,
            caller_service="gateway",
            scopes=["agent:read", "agent:write"],
        )
    return _agent_client


def _get_inference_client() -> InternalClient:
    global _inference_client
    if _inference_client is None:
        _inference_client = InternalClient(
            base_url=_INFERENCE_URL,
            caller_service="gateway",
            scopes=["inference:read", "inference:write"],
        )
        _inference_client.client.timeout = httpx.Timeout(_INFERENCE_TIMEOUT_S)
    return _inference_client


# ── Helpers ───────────────────────────────────────────────────────────────────


def _safe_headers(request: Request) -> Dict[str, str]:
    """Pass-through headers that are safe to forward downstream."""
    forwarded: Dict[str, str] = {}
    for hdr in ("X-Trace-Id", "X-Request-Id", "Content-Type", "Idempotency-Key"):
        val = request.headers.get(hdr)
        if val:
            forwarded[hdr] = val
    return forwarded


async def _proxy(
    client: InternalClient,
    method: str,
    path: str,
    ctx: AuthContext,
    request: Request,
    body: Optional[bytes] = None,
) -> JSONResponse:
    """Execute a proxied request and map errors to appropriate HTTP responses."""
    extra_headers = _safe_headers(request)
    kwargs: Dict[str, Any] = {"headers": extra_headers}
    if body:
        kwargs["content"] = body

    try:
        resp = await client._request(method, path, ctx, **kwargs)
        # Forward the downstream status code and JSON body unchanged.
        try:
            return JSONResponse(content=resp.json(), status_code=resp.status_code)
        except Exception:
            return JSONResponse(content={"raw": resp.text}, status_code=resp.status_code)

    except CircuitBreakerOpenError:
        logger.warning("Circuit breaker open for path=%s", path)
        raise HTTPException(
            status_code=503,
            detail={
                "code": "SERVICE_UNAVAILABLE",
                "message": "Downstream service is temporarily unavailable.",
                "retryable": True,
            },
        )
    except InternalClientError as exc:
        raise HTTPException(
            status_code=exc.status_code,
            detail={
                "code": exc.error.code,
                "message": exc.error.message,
                "retryable": exc.error.retryable,
            },
        )
    except (httpx.TimeoutException, httpx.NetworkError) as exc:
        logger.error("Network error proxying %s %s: %s", method, path, exc)
        raise HTTPException(
            status_code=503,
            detail={
                "code": "GATEWAY_TIMEOUT",
                "message": "Upstream service did not respond in time.",
                "retryable": True,
            },
        )


# ── Mission proxy routes ──────────────────────────────────────────────────────


@router.get("/api/v1/missions")
async def proxy_list_missions(
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.VIEWER)),
):
    """Proxy GET /api/v1/missions → Mission service."""
    qs = str(request.url.query)
    path = "/api/v1/missions" + (f"?{qs}" if qs else "")
    return await _proxy(_get_mission_client(), "GET", path, ctx, request)


@router.post("/api/v1/missions", status_code=201)
async def proxy_create_mission(
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.ANALYST)),
):
    """Proxy POST /api/v1/missions → Mission service."""
    body = await request.body()
    return await _proxy(_get_mission_client(), "POST", "/api/v1/missions", ctx, request, body)


@router.get("/api/v1/missions/{mission_id}")
async def proxy_get_mission(
    mission_id: str,
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.VIEWER)),
):
    """Proxy GET /api/v1/missions/{id} → Mission service."""
    return await _proxy(
        _get_mission_client(), "GET", f"/api/v1/missions/{mission_id}", ctx, request
    )


@router.patch("/api/v1/missions/{mission_id}")
async def proxy_update_mission(
    mission_id: str,
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.ANALYST)),
):
    """Proxy PATCH /api/v1/missions/{id} → Mission service."""
    body = await request.body()
    return await _proxy(
        _get_mission_client(), "PATCH", f"/api/v1/missions/{mission_id}", ctx, request, body
    )


@router.delete("/api/v1/missions/{mission_id}", status_code=204)
async def proxy_delete_mission(
    mission_id: str,
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.OPERATOR)),
):
    """Proxy DELETE /api/v1/missions/{id} → Mission service (OPERATOR+)."""
    return await _proxy(
        _get_mission_client(), "DELETE", f"/api/v1/missions/{mission_id}", ctx, request
    )


@router.post("/api/v1/missions/{mission_id}/runs", status_code=202)
async def proxy_submit_run(
    mission_id: str,
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.OPERATOR)),
):
    """Proxy POST /api/v1/missions/{id}/runs → Mission service (OPERATOR+)."""
    body = await request.body()
    return await _proxy(
        _get_mission_client(),
        "POST",
        f"/api/v1/missions/{mission_id}/runs",
        ctx,
        request,
        body,
    )


@router.get("/api/v1/jobs/{job_id}")
async def proxy_get_job(
    job_id: str,
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.VIEWER)),
):
    """Proxy GET /api/v1/jobs/{job_id} → Mission service."""
    return await _proxy(_get_mission_client(), "GET", f"/api/v1/jobs/{job_id}", ctx, request)


# ── Agent proxy routes ────────────────────────────────────────────────────────


@router.post("/api/v1/agent/plan")
async def proxy_agent_plan(
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.ANALYST)),
):
    """Proxy POST /api/v1/agent/plan → Agent service (ANALYST+)."""
    body = await request.body()
    return await _proxy(_get_agent_client(), "POST", "/api/v1/agent/plan", ctx, request, body)


@router.post("/api/v1/agent/execute", status_code=202)
async def proxy_agent_execute(
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.ANALYST)),
):
    """Proxy POST /api/v1/agent/execute → Agent service (ANALYST+)."""
    body = await request.body()
    return await _proxy(_get_agent_client(), "POST", "/api/v1/agent/execute", ctx, request, body)


@router.get("/api/v1/agent/runs/{job_id}")
async def proxy_agent_run_status(
    job_id: str,
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.VIEWER)),
):
    """Proxy GET /api/v1/agent/runs/{job_id} → Agent service."""
    return await _proxy(_get_agent_client(), "GET", f"/api/v1/agent/runs/{job_id}", ctx, request)


@router.post("/api/v1/agent/sensor-decision")
async def proxy_agent_sensor_decision(
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.ANALYST)),
):
    """Proxy POST /api/v1/agent/sensor-decision → Agent service."""
    body = await request.body()
    return await _proxy(
        _get_agent_client(), "POST", "/api/v1/agent/sensor-decision", ctx, request, body
    )


@router.post("/api/v1/agent/confidence")
async def proxy_agent_confidence(
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.ANALYST)),
):
    """Proxy POST /api/v1/agent/confidence → Agent service."""
    body = await request.body()
    return await _proxy(_get_agent_client(), "POST", "/api/v1/agent/confidence", ctx, request, body)


@router.get("/api/v1/agent/tools")
async def proxy_agent_tools(
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.VIEWER)),
):
    """Proxy GET /api/v1/agent/tools → Agent service."""
    return await _proxy(_get_agent_client(), "GET", "/api/v1/agent/tools", ctx, request)


# ── Inference proxy routes (P3) ──────────────────────────────────────────────
#
# Without these the browser could not reach the model at all: it may only call the
# gateway, and the gateway had no route to the inference service. The roles match
# the inference service's own -- ANALYST to run an analysis, VIEWER to read -- so
# the user's role is enforced here, and inference trusts the gateway's S2S token.


@router.post("/api/v1/inference/analyses")
async def proxy_inference_analyse(
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.ANALYST)),
):
    """Proxy POST /api/v1/inference/analyses → Inference service (ANALYST+).

    Returns ``Analysis | Abstention`` unchanged. Both are 200: an abstention is a
    considered answer, not a transport failure.
    """
    body = await request.body()
    return await _proxy(
        _get_inference_client(), "POST", "/api/v1/inference/analyses", ctx, request, body
    )


@router.get("/api/v1/inference/models")
async def proxy_inference_models(
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.VIEWER)),
):
    """Proxy GET /api/v1/inference/models → Inference service.

    The registry: each model's held-out score, calibration, and whether its
    checkpoint has been verified against its committed manifest.
    """
    return await _proxy(_get_inference_client(), "GET", "/api/v1/inference/models", ctx, request)


@router.get("/api/v1/inference/analyses/{trace_id}/extent")
async def proxy_inference_extent(
    trace_id: uuid.UUID,
    request: Request,
    ctx: AuthContext = Depends(require_role(Role.VIEWER)),
):
    """Proxy GET .../analyses/{trace_id}/extent → Inference service.

    GeoJSON, not JSON-wrapped: the map consumes it directly. ``trace_id`` is parsed
    as a UUID here as well as downstream, so the path forwarded is always one this
    gateway built -- A10 above -- never a caller's string.
    """
    path = f"/api/v1/inference/analyses/{trace_id}/extent"
    client = _get_inference_client()
    try:
        response = await client._request("GET", path, ctx, headers=_safe_headers(request))
    except CircuitBreakerOpenError:
        raise HTTPException(
            status_code=503,
            detail={
                "code": "SERVICE_UNAVAILABLE",
                "message": "Inference service is temporarily unavailable.",
                "retryable": True,
            },
        )
    except InternalClientError as exc:
        raise HTTPException(
            status_code=exc.status_code,
            detail={
                "code": exc.error.code,
                "message": exc.error.message,
                "retryable": exc.error.retryable,
            },
        )
    except (httpx.TimeoutException, httpx.NetworkError):
        raise HTTPException(
            status_code=503,
            detail={
                "code": "GATEWAY_TIMEOUT",
                "message": "Inference service did not respond in time.",
                "retryable": True,
            },
        )
    return Response(content=response.content, media_type="application/geo+json")